+++ .dockerignore
@@ -11,3 +11,4 @@ __pycache__
 
 .coverage
 .vscode
+.ipynb_checkpoints
