--- data/fafb.yml
@@ -1,21 +1,0 @@
-data_source:
-  EDGES: "gs://chunked-graph/190410_FAFB_v02/edges"
-  COMPONENTS: "gs://chunked-graph/190410_FAFB_v02/components"
-  WATERSHED: "gs://seunglab2/drosophila_v0/ws_190410_FAFB_v02_ws_size_threshold_200"
-  DATA_VERSION: 2
-
-graph_config:
-  CHUNK_SIZE: [256, 256, 512]
-  FANOUT: 2
-  SPATIAL_BITS: 10
-  ID_PREFIX: "aa-"
-
-backend_client:
-  TYPE: "bigtable"
-  CONFIG:
-    ADMIN: true
-    READ_ONLY: false
-
-ingest_config:
-  AGGLOMERATION: "gs://ranl-scratch/190410_FAFB_v02_ws_size_threshold_200/agg"
-  WATERSHED: "gs://seunglab2/drosophila_v0/ws_190410_FAFB_v02_ws_size_threshold_200"
