--- data/minnie65.yml
@@ -1,21 +1,0 @@
-data_source:
-  EDGES: "gs://chunked-graph/minnie65_0/edges"
-  COMPONENTS: "gs://chunked-graph/minnie65_0/components"
-  WATERSHED: "gs://microns-seunglab/minnie65/ws_minnie65_0"
-  DATA_VERSION: 2
-
-graph_config:
-  CHUNK_SIZE: [256, 256, 512]
-  FANOUT: 2
-  SPATIAL_BITS: 10
-  ID_PREFIX: ""
-
-backend_client:
-  TYPE: "bigtable"
-  CONFIG:
-    ADMIN: true
-    READ_ONLY: false
-
-ingest_config:
-  AGGLOMERATION: "gs://ranl-scratch/minnie65_0/agg"
-  WATERSHED: "gs://microns-seunglab/minnie65/ws_minnie65_0"
