--- data/minnie65_phase3.yml
@@ -1,21 +1,0 @@
-data_source:
-  EDGES: "gs://chunked-graph/minnie_p3/edges"
-  COMPONENTS: "gs://chunked-graph/minnie_p3/components"
-  WATERSHED: "gs://ranl-minnie-ws/neuroglancer/minnie_p3"
-  DATA_VERSION: 4
-
-graph_config:
-  CHUNK_SIZE: [256, 256, 512]
-  FANOUT: 2
-  SPATIAL_BITS: 10
-  ID_PREFIX: ""
-
-backend_client:
-  TYPE: "bigtable"
-  CONFIG:
-    ADMIN: true
-    READ_ONLY: false
-
-ingest_config:
-  AGGLOMERATION: "gs://ranl-minnie-rg/minnie_p3/agg"
-  WATERSHED: "gs://ranl-minnie-ws/neuroglancer/minnie_p3"
